+++ lib/bulbs-elements/util/load-on-demand.js
@@ -29,7 +29,10 @@ function maybeLoadComponents () {
   if (animationFrameRequested === false) {
     requestAnimationFrame(() => {
       animationFrameRequested = false;
-      pendingComponents.forEach(maybeLoadComponent);
+      pendingComponents
+        .map(component => [component, component.getBoundingClientRect()])
+        .forEach(maybeLoadComponent)
+      ;
     });
     animationFrameRequested = true;
   }
@@ -128,26 +131,64 @@ export class LoadOnDemand extends React.Component {
 }
 
 export default function loadOnDemand (Component) {
-  class Wrapped extends React.Component {
-    static displayName = `LoadOnDemand(${Component.displayName}`;
-    static propTypes = {
-      disableLazyLoading: PropTypes.string,
-    };
-
-    render () {
-      if (typeof this.props.disableLazyLoading !== 'undefined') {
-        return <Component {...this.props} />;
+  if (Component.prototype instanceof HTMLElement) {
+    return class Wrapped extends Component {
+      setState (newState) {
+        Object.extend(this.state, newState);
       }
 
-      return (
-        <LoadOnDemand
-          ref="loadOnDemand"
-          component={Component}
-          componentProps={this.props}
-        />
-      );
-    }
-  }
+      createdCallback () {
+        this.state = { loaded: false };
+        if (super.createdCallback) {
+          super.createdCallback();
+        }
+      }
 
-  return Wrapped;
+      attachedCallback () {
+        //if (!this.state.loaded) {
+        //  enqueue(this);
+        //}
+        if (super.attachedCallback) {
+          super.attachedCallback();
+        }
+      }
+
+      detachedCallback () {
+        //if (this.state.loaded) {
+        //  dequeue(this);
+        //}
+        if (super.detachedCallback) {
+          super.detachedCallback();
+        }
+      }
+
+      attributeChangedCallback (attrName, newValue, oldValue) {
+        if (super.attributeChangedCallback) {
+          super.attributeChangedCallback(attrName, newValue, oldValue);
+        }
+      }
+    };
+  }
+  else { // eslint-disable-line no-else-return
+    return class Wrapped extends React.Component {
+      static displayName = `LoadOnDemand(${Component.displayName}`;
+      static propTypes = {
+        disableLazyLoading: PropTypes.string,
+      };
+
+      render () {
+        if (typeof this.props.disableLazyLoading !== 'undefined') {
+          return <Component {...this.props} />;
+        }
+
+        return (
+          <LoadOnDemand
+            ref="loadOnDemand"
+            component={Component}
+            componentProps={this.props}
+          />
+        );
+      }
+    };
+  }
 }
